Use a randomized namespace for api extensions test
This change uses a randomized namespace for the api extensions test. It
uses the go templating library to template the namespace into the
fixture files. This change also moves the fixture files to be relative
to the test, so that they are easier to reference.

This change also properly waits on the exit code for the kubectl
commands when cleaning up. Lastly, this change adds a
DefaultEventuallyTimeout of 60s for the api_extensions_test, as we were
setting that as the timeout for every Eventually block.

[#153565043]

Signed-off-by: James Myers <jmyers@pivotal.io>