Updated integration, turbulence, and conformance tests to use a config
Rather than utilizing a slew of ENV variables, Ginkgo can be run with
just a single JSON config file. This config can be generated using
`scripts/generate-test-config.sh` while logged in to BOSH and supplied
to Ginkgo by setting the ENV variable `CONFIG`. For example:

```
CONFIG=<(./scripts/generate-test-config.sh) ginkgo src/tests/integration-tests/generic
```

Additionally, the scripts `run-k8s-conformance-tests.sh`,
`run-k8s-integration-tests.sh` and `run-k8s-turbulence-tests.sh` can be
run in a local dev environment, passing along an environment and
deployment name. Use the `-h` flag to view the list of arguments
required for each script.

The pipeline now has its own suite of scripts in `scripts/ci/` that
configure and invoke their respective test-scripts. It's the hope that
we can migrate the CI-specific test scripts into this directory while
leaving locally-runnable copies (where sensible) in the `scripts/`
directory. This should help to make it more clear which scripts are
runnable during the development process and which are intended to run in
the CI.

[#152770162]

Signed-off-by: Travis Hall <thall@vmware.com>