Upped the max_in_flight number for AWS multi-az worker deployment
This needs to be explicitly done since the first step of AWS deployment
is just deploying the master (and route-sync). Post the changes that
need to be included in the manifest, the deploy_k8s binary treats the
addition of workers as updating a complete deployment which is not true.
Hence bumping the max in flight to 100% to update multiple worker nodes
in parallel.