Shorten retry on VM-start detection in failure tests
We think that it may missing the window during which a VM claims to be
starting and therefore never detecting that the VM is anything but
'running'.

BOSH should probably have a different state for the immediate recovery
state (e.g. 'recovering') rather than 'running' so that it doesn't go
'running' -> 'starting' -> 'running', making detection much more
difficult.