Not fail tests if security group deletion fails
We had the logic to clean up security group to overcome this https://github.com/kubernetes/kubernetes/issues/17626
We added the tag to the worker security group and now the security group
for load balancer got deleted successfully. This means we don't have to
clean it up. And when the rule is not there, it shuold be ignored.

We have added extra output to check if security group got actually
removed by Kubernetes. We have also added a chore to remove this logic
completely here: https://www.pivotaltracker.com/story/show/157658693

[#157492449]

Signed-off-by: Akshay Mankar <amankar@pivotal.io>