Create workaround for not using gcs-bosh-creds
- The turbulence tests don't need gcs-bosh-creds, it can pull all the
info from the lock file for the test config. However generate test
config sources a utils scripts that expects that file. Untagling this
from the utils script is a lot of effort and changes a lot of scripts
relying on it hence a workaround is being used to check the existence of
gcs-bosh-creds

[#159894898]