Add a new step to download dependencies
The underlying problem here is that bosh will over-optimize when
given a command like

      bosh upload-release <remote URL>

It will refuse to upload if the director already has a release with the
same name/version, even though the remote URL was compiled with a
different stem cell.

However, a command like

      bosh upload-release ./path

will never be optimized, and mulitple releases for the same (upstream)
version, compiled for *different* stemcells, can coexist on the director
and be correctly used later for the specified stemcell.

Therefore, this "manual" download step, when combined with a "put" step
that identifies the directory
of the downloaded releases, will remove a whole class of pipeline "flakes"
when the director already has a compiled release for an old stemcell,
but a new stemcell is introduced, along with all the releases that have
been compiled for the new stemcell.